[MINT-2208] Homepage refactor (#1245)
* Changed file names, minor refactor

* Fixed route and models favorites props

* Remove feature flag

* Targeted first table in cypress

* Merge in main, resolved conflicts, move HomeTable

* Removed unneeded flag

* Updated snapshots
