+++ src/views/Home/Settings/selectSolutions.tsx
