+++ src/views/Home/components/ModelsBySolutions/banner.tsx
@@ -6,7 +6,7 @@ import { ModelBySolutionStatus, OperationalSolutionKey } from 'gql/gen/graphql';
 
 import useCheckResponsiveScreen from 'hooks/useCheckMobile';
 import { helpSolutions } from 'views/HelpAndKnowledge/SolutionsHelp/solutionsMap';
-import { ModelsBySolutionType } from 'views/HomeNew/components/ModelsBySolutions/table';
+import { ModelsBySolutionType } from 'views/Home/components/ModelsBySolutions/table';
 
 import './index.scss';
