+++ src/views/Home/components/ModelsBySolutions/card.tsx
