+++ src/views/Home/components/ModelsBySolutions/index.scss
