+++ src/views/Home/index.scss
@@ -1,7 +1,16 @@
 @use 'uswds-core' as *;
 
 .home {
-  &__hr {
-    border-top: 1px solid color('base-light');
-  }
+    &__card {
+        flex-direction: row-reverse;
+
+        .usa-card__container {
+            padding: 1rem;
+            border-radius: .25rem;
+            border: 2px solid color('base-lighter');    
+            box-shadow: 0 4px 8px 0px rgba(0, 0, 0, 0.1);
+            height: auto !important;
+            margin: 0 !important;
+        }
+    }
 }
