--- src/views/HomeNew/index.scss
@@ -1,14 +1,0 @@
-@use 'uswds-core' as *;
-
-.home {
-    &__card {
-        .usa-card__container {
-            padding: 1rem;
-            border-radius: .25rem;
-            border: 2px solid color('base-lighter');    
-            box-shadow: 0 4px 8px 0px rgba(0, 0, 0, 0.1);
-            height: auto !important;
-            margin: 0 !important;
-        }
-    }
-}
