+++ src/views/ModelPlan/ModelPlanOverview/index.tsx
@@ -12,7 +12,8 @@ import {
 import {
   GetFavoritesQuery,
   ModelPlanFilter,
-  useGetFavoritesQuery
+  useGetFavoritesQuery,
+  ViewCustomizationType
 } from 'gql/gen/graphql';
 
 import FavoritesTable from 'components/FavoriteCard/table';
@@ -27,7 +28,7 @@ import useMessage from 'hooks/useMessage';
 import { AppState } from 'reducers/rootReducer';
 import { isMAC } from 'utils/user';
 
-import ModelPlansTable from '../Table';
+import ModelPlansTable from '../../Home/Table';
 
 type GetFavoritesModelPlanCollection = GetFavoritesQuery['modelPlanCollection'][0];
 
@@ -35,7 +36,7 @@ export type UpdateFavoriteProps = 'addFavorite' | 'removeFavorite';
 
 const ModelPlan = () => {
   const { t } = useTranslation('readOnlyModelPlan');
-  const { t: h } = useTranslation('home');
+  const { t: h } = useTranslation('customHome');
 
   const userGroups = useSelector((state: AppState) => state.auth.groups);
   const macUser = isMAC(userGroups);
@@ -156,7 +157,9 @@ const ModelPlan = () => {
           </p>
 
           <ModelPlansTable
-            type="models"
+            id="favorite-model-plans-table"
+            type={ViewCustomizationType.ALL_MODEL_PLANS}
+            isHome={false}
             updateFavorite={handleUpdateFavorite}
           />
         </Grid>
