--- src/views/ModelPlan/Table/index.scss
@@ -1,6 +1,0 @@
-.model-plan-table {
-  th, td {
-    position: relative;
-    word-break: break-word;
-  }
-}
